+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # **CHANGELOG**
 
+## **1.0.2**
+### **CHANGES**
+- Renamed master to management as part of "Words Matter Initiative"
+
+### **BUG FIXES**
+- Fixed bug related to metadata files not getting copied from management host to worker nodes
+
 ## **1.0.1**
-###CHANGES
-Changes to post provisioning scripts to mitigate Polkit Local Privilege Escalation Vulnerability (CVE-2021-4034).
+### **CHANGES**
+- Changes to post provisioning scripts to mitigate Polkit Local Privilege Escalation Vulnerability (CVE-2021-4034).
 
 ## **1.0.0**
 - Initial Release